+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/elementwise/SubtractOp.java
@@ -52,7 +52,7 @@ public class SubtractOp extends BaseTwoArrayElementWiseOp {
             if(origin instanceof IComplexNDArray) {
                 IComplexNDArray originComplexValue = (IComplexNDArray) origin;
                 IComplexNumber firstValue = (IComplexNumber) originComplexValue.element();
-                double realValue = (double) value.element();
+                double realValue = (float) value.element();
                 IComplexNumber retValue = firstValue.sub(realValue);
                 return NDArrays.scalar(retValue);
 
@@ -60,8 +60,8 @@ public class SubtractOp extends BaseTwoArrayElementWiseOp {
 
             //both normal
             else {
-                double firstElement = (double) origin.element();
-                double secondElement = (double) value.element();
+                float firstElement = (float) origin.element();
+                float secondElement = (float) value.element();
                 return NDArrays.scalar(firstElement - secondElement,0);
             }
